+++ wagtailtinymce/rich_text.py
@@ -27,12 +27,20 @@ from __future__ import absolute_import, unicode_literals
 
 import json
 
+from django.conf import settings
 from django.forms import widgets
 from django.utils import translation
+from wagtail import __version__ as WAGTAIL_VERSION
 from wagtail.utils.widgets import WidgetWithScript
-from wagtail.wagtailadmin.edit_handlers import RichTextFieldPanel
-from wagtail.wagtailcore.rich_text import DbWhitelister
-from wagtail.wagtailcore.rich_text import expand_db_html
+
+if WAGTAIL_VERSION >= '2.0':
+    from wagtail.admin.edit_handlers import RichTextFieldPanel
+    from wagtail.admin.rich_text.converters.editor_html import \
+        EditorHTMLConverter
+    from wagtail.core.rich_text import features
+else:
+    from wagtail.wagtailadmin.edit_handlers import RichTextFieldPanel
+    from wagtail.wagtailcore.rich_text import DbWhitelister, expand_db_html
 
 
 class TinyMCERichTextArea(WidgetWithScript, widgets.Textarea):
@@ -56,7 +64,8 @@ class TinyMCERichTextArea(WidgetWithScript, widgets.Textarea):
             'options': {
                 'browser_spellcheck': True,
                 'noneditable_leave_contenteditable': True,
-                'language': translation.to_locale(translation.get_language()),
+                'language': translation.to_locale(
+                    translation.get_language() or settings.LANGUAGE_CODE),
                 'language_load': True,
             },
         }
@@ -64,18 +73,29 @@ class TinyMCERichTextArea(WidgetWithScript, widgets.Textarea):
     def __init__(self, attrs=None, **kwargs):
         super(TinyMCERichTextArea, self).__init__(attrs)
         self.kwargs = self.getDefaultArgs()
+        self.features = kwargs.pop('features', None)
         if kwargs is not None:
             self.kwargs.update(kwargs)
+
+        if WAGTAIL_VERSION >= '2.0':
+            if self.features is None:
+                self.features = features.get_default_features()
+                self.converter = EditorHTMLConverter()
+            else:
+                self.converter = EditorHTMLConverter(self.features)
 
     def get_panel(self):
         return RichTextFieldPanel
 
-    def render(self, name, value, attrs=None):
+    def render(self, name, value, attrs=None, renderer=None):
         if value is None:
             translated_value = None
         else:
-            translated_value = expand_db_html(value, for_editor=True)
-        return super(TinyMCERichTextArea, self).render(name, translated_value, attrs)
+            if WAGTAIL_VERSION >= '2.0':
+                translated_value = self.converter.from_database_format(value)
+            else:
+                translated_value = expand_db_html(value, for_editor=True)
+        return super(TinyMCERichTextArea, self).render(name, translated_value, attrs, renderer)
 
     def render_js_init(self, id_, name, value):
         kwargs = {
@@ -97,10 +117,30 @@ class TinyMCERichTextArea(WidgetWithScript, widgets.Textarea):
             else:
                 kwargs['menubar'] = ' '.join(self.kwargs['menus'])
 
+        if 'passthru_init_keys' in self.kwargs:
+            kwargs.update(self.kwargs['passthru_init_keys'])
+
+        if 'table' in self.kwargs:
+            for key, values in self.kwargs['table'].items():
+                kwargs[f'table_{key}'] = values
+
         return "makeTinyMCEEditable({0}, {1});".format(json.dumps(id_), json.dumps(kwargs))
 
     def value_from_datadict(self, data, files, name):
         original_value = super(TinyMCERichTextArea, self).value_from_datadict(data, files, name)
         if original_value is None:
             return None
-        return DbWhitelister.clean(original_value)
+        if WAGTAIL_VERSION >= '2.0':
+            return self.converter.to_database_format(original_value)
+        else:
+            return DbWhitelister.clean(original_value)
+
+    def format_value(self, value):
+        # Convert database rich text representation to the format required by
+        # the input field
+        value = super().format_value(value)
+
+        if value is None:
+            value = ""
+
+        return self.converter.from_database_format(value)